Fixing : Getting Current Location

+++ app/src/main/java/com/riantono/weather/repository/entity/Main.kt
@@ -6,5 +6,5 @@ import kotlinx.android.parcel.Parcelize
 
 @Parcelize
 data class Main(@field:SerializedName("temp") val temp: Double,
-                @field:SerializedName("pressure") val pressure: Int,
-                @field:SerializedName("humidity") val humidity: Int) : Parcelable
+                @field:SerializedName("pressure") val pressure: Double,
+                @field:SerializedName("humidity") val humidity: Double) : Parcelable
