+++ app/src/main/java/com/riantono/weather/ui/entity/WeatherModel.kt
@@ -4,4 +4,4 @@ import android.os.Parcelable
 import kotlinx.android.parcel.Parcelize
 
 @Parcelize
-data class WeatherModel(val temp: Double, val humidity: Int, val windSpeed: Double, val description: String?) : Parcelable
+data class WeatherModel(val temp: Double, val humidity: Double, val windSpeed: Double, val description: String?) : Parcelable
